feat(ci): allow shippable ci to build this project
