+++ .gitignore
@@ -197,3 +197,4 @@ FakesAssemblies/
 
 # VIM files
 .*.swp
+Amazon.SQS.ExtendedClient.v3.ncrunchsolution
