+++ src/Amazon.SQS.ExtendedClient.Core/Amazon.SQS.ExtendedClient.v3.ncrunchproject
