+++ src/Amazon.SQS.ExtendedClient.Core/AmazonSQSExtendedClientBase.cs
