+++ src/Amazon.SQS.ExtendedClient.Core/ExtendedClientConfiguration.cs
