+++ src/Amazon.SQS.ExtendedClient.Core/GuidS3KeyProvider.cs
@@ -1,0 +1,12 @@
+﻿namespace Amazon.SQS.ExtendedClient
+{
+    using System;
+
+    public class GuidS3KeyProvider : IS3KeyPovider
+    {
+        public string GenerateName()
+        {
+            return Guid.NewGuid().ToString("N");
+        }
+    }
+}
