+++ src/Amazon.SQS.ExtendedClient.Core/IS3KeyPovider.cs
@@ -1,0 +1,7 @@
+﻿namespace Amazon.SQS.ExtendedClient
+{
+    public interface IS3KeyPovider
+    {
+        string GenerateName();
+    }
+}
