+++ src/Amazon.SQS.ExtendedClient.Core/MessageS3Pointer.cs
