+++ src/Amazon.SQS.ExtendedClient.Core/Properties/AssemblyInfo.cs
