+++ src/Amazon.SQS.ExtendedClient.Core/SQSExtendedClientConstants.cs
