+++ src/Amazon.SQS.ExtendedClient.Core/packages.config
