--- src/Amazon.SQS.ExtendedClient/Amazon.SQS.ExtendedClient.nuspec
@@ -1,24 +1,0 @@
-<?xml version="1.0"?>
-<package >
-  <metadata>
-    <id>Amazon.SQS.ExtendedClient</id>
-    <version>$version$</version>
-    <title>Amazon.SQS.ExtendedClient</title>
-    <authors>Oleg Rakitskiy</authors>
-    <owners>Oleg Rakitskiy</owners>
-    <licenseUrl>https://github.com/raol/amazon-sqs-net-extended-client-lib/blob/master/LICENSE</licenseUrl>
-    <projectUrl>https://github.com/raol/amazon-sqs-net-extended-client-lib</projectUrl>
-    <requireLicenseAcceptance>false</requireLicenseAcceptance>
-    <description>Extension to Amazon SQS that adds support for sending and receiving messages greater than 256K</description>
-    <copyright>Copyright 2016</copyright>
-    <tags>AWS Amazon SQS cloud</tags>
-    <dependencies>
-      <dependency id="AWSSDK.S3" version="3.3" />
-      <dependency id="AWSSDK.SQS" version="3.3" />
-      <dependency id="Newtonsoft.Json" version="10.0" />
-    </dependencies>
-  </metadata>
-  <files>
-    <file src="bin\Release\Amazon.SQS.ExtendedClient.dll" target="lib\net45" />
-  </files>
-</package>
