--- src/Amazon.SQS.ExtendedClient/IS3KeyPovider.cs
@@ -1,7 +1,0 @@
-﻿namespace Amazon.SQS.ExtendedClient
-{
-    public interface IS3KeyPovider
-    {
-        string GenerateName();
-    }
-}
